+++ bpm-engine-api/src/main/java/io/takari/bpm/model/CallActivity.java
@@ -10,16 +10,26 @@ public class CallActivity extends AbstractElement {
     private final String calledElement;
     private final Set<VariableMapping> in;
     private final Set<VariableMapping> out;
+    private final boolean copyAllVariables;
     
     public CallActivity(String id, String calledElement) {
-        this(id, calledElement, null, null);
+        this(id, calledElement, null, null, false);
+    }
+
+    public CallActivity(String id, String calledElement, boolean copyAllVariables) {
+        this(id, calledElement, null, null, copyAllVariables);
     }
 
     public CallActivity(String id, String calledElement, Set<VariableMapping> in, Set<VariableMapping> out) {
+        this(id, calledElement, in, out, false);
+    }
+
+    public CallActivity(String id, String calledElement, Set<VariableMapping> in, Set<VariableMapping> out, boolean copyAllVariables) {
         super(id);
         this.calledElement = calledElement;
         this.in = in;
         this.out = out;
+        this.copyAllVariables = copyAllVariables;
     }
 
     public String getCalledElement() {
@@ -40,5 +50,9 @@ public class CallActivity extends AbstractElement {
 
     public void setName(String name) {
         this.name = name;
+    }
+
+    public boolean isCopyAllVariables() {
+        return copyAllVariables;
     }
 }