+++ bpm-engine-impl/src/main/java/io/takari/bpm/actions/MakeSubProcessVariablesAction.java
@@ -9,17 +9,23 @@ public class MakeSubProcessVariablesAction implements Action {
     private static final long serialVersionUID = 1L;
 
     private final Set<VariableMapping> inVariables;
+    private final boolean copyAllVariables;
 
-    public MakeSubProcessVariablesAction(Set<VariableMapping> inVariables) {
+    public MakeSubProcessVariablesAction(Set<VariableMapping> inVariables, boolean copyAllVariables) {
         this.inVariables = inVariables;
+        this.copyAllVariables = copyAllVariables;
     }
 
     public Set<VariableMapping> getInVariables() {
         return inVariables;
     }
 
+    public boolean isCopyAllVariables() {
+        return copyAllVariables;
+    }
+
     @Override
     public String toString() {
-        return "MakeSubProcessVariablesAction [inVariables=" + inVariables + "]";
+        return "MakeSubProcessVariablesAction [inVariables=" + inVariables + ", copyAllVariables=" + copyAllVariables + "]";
     }
 }